+++ src/library/Hasql/Session.hs
@@ -14,11 +14,11 @@ module Hasql.Session
 where
 
 import Core.Contexts.ParamsEncoder qualified as ParamsEncoder
-import Core.Contexts.ResultConsumer qualified as ResultConsumer
 import Core.Contexts.Session qualified as Session
 import Core.Errors
 import Hasql.Connection qualified as Connection
 import Hasql.Statement qualified as Statement
+import Hipq.ResultDecoder qualified as ResultDecoder
 import Platform.Prelude
 
 -- |
@@ -32,12 +32,10 @@ run session connection = Connection.use connection session
 -- |
 -- Execute a statement by providing parameters to it.
 statement :: params -> Statement.Statement params result -> Session.Session result
-statement
-  params
-  ( Statement.Statement
-      sql
-      (ParamsEncoder.unwrap -> paramsEncoder)
-      (ResultConsumer.unwrap -> decoder)
-      preparable
-    ) =
-    Session.statement sql paramsEncoder decoder preparable params
+statement params (Statement.Statement sql paramsEncoder decoder preparable) =
+  Session.statement
+    sql
+    (ParamsEncoder.unwrap paramsEncoder)
+    (ResultDecoder.unwrap decoder)
+    preparable
+    params